Update README.md
The primary objective of this data science project is to develop a robust classification system for farm insects, with special attention to identifying and differentiating the specific pests responsible for orange canker. By leveraging advanced machine learning algorithms and data analytics, we aim to create a predictive model that can accurately classify insect species based on their characteristics, behaviors, and potential impact on crops.